feat: новая верстка слайдера
