+++ src/components/Slider/Slider.scss
@@ -2,15 +2,13 @@
 @import '../../styles/abstract/mixins.scss';
 
 .swiper__wrapper {
-  
+  position: relative;
 }
 
 .swiper {
+  position: unset !important;
   margin-top: 74px;
-  // padding: 0 80px !important;
   height: 310px;
-  overflow-x: clip !important;
-  overflow-y: unset !important;
 }
 
 .swiper-slide {
@@ -30,6 +28,7 @@
   }
 
   &__img-container {
+    width: 422px;
     flex-grow: 1;
   }
 
@@ -42,21 +41,15 @@
 
   &__text-container {
     display: flex;
+    gap: 15px;
     flex-direction: column;
     justify-content: space-between;
-    // gap: 60px;
-    // position: absolute;
-    // top: 0%;
-    // bottom: 0%;
-    // left: 0%;
-    // right: 40%;
-    // z-index: 5;
     width: 510px;
-    padding: 50px 60px;
-    // overflow-wrap: anywhere;
+    padding: 50px 60px 40px;
   }
 
   &__text {
+    overflow: hidden;
     margin: 0;
     color: $white-standart;
     text-align: start;
@@ -68,7 +61,7 @@
   &__btn {
     display: flex;
     width: 284px;
-    height: 55px;
+    min-height: 55px;
     background-color: $accent-lime-button;
     border-radius: 32px;
     text-decoration: none;
@@ -81,7 +74,7 @@
 }
 
 .swiper-pagination {
-  bottom: -33px !important;
+  bottom: -28px !important;
 }
 
 .swiper-pagination-bullet {
@@ -102,16 +95,17 @@
   color: $dark-grey;
   border: 1px solid $border-input;
   z-index: 1;
-  top: 130px;
+  top: 120px;
+  cursor: pointer;
 }
 
 .swiper-button-prev {
-  left: 1px;
+  left: -23px;
   background-image: url('../../images/slider-nav-chevron.svg');
 }
 
 .swiper-button-next {
-  right: 0px;
+  right: -23px;
   background-image: url('../../images/slider-nav-chevron.svg');
   transform: rotate(180deg);
 }
