+++ src/components/Slider/SliderTypes.tsx
@@ -1,0 +1,9 @@
+type Slider = {
+  text: string;
+  img: string;
+  link: string;
+}[];
+
+type SliderProps = {
+  slides: Slider;
+};
