+++ src/pages/HomePage/HomePage.tsx
@@ -3,12 +3,13 @@ import Catalogue from '../../components/Catalogue/Catalogue';
 import Slider from '../../components/Slider/Slider';
 import Recommended from '../../components/Recommended/Recommended';
 import { Categories } from '../../components/Categories/Categories';
+import { slides } from '../../utils/constants';
 
 export const HomePage: FC = () => {
     return (
         <>
         <Categories />
-        <Slider />
+        <Slider slides={slides}/>
         <Catalogue />
         <Recommended />
         </>
